Testing Conversation and Friendship controllers and services.

+++ Vänskap-Api/Service/FriendshipService.cs
@@ -3,6 +3,7 @@ using System.Reflection.Metadata.Ecma335;
 using System.Security.Claims;
 using Vänskap_Api.Data;
 using Vänskap_Api.Models;
+using Vänskap_Api.Models.Dtos.Friend;
 using Vänskap_Api.Models.Dtos.User;
 using Vänskap_Api.Service.IService;
 
@@ -42,32 +43,42 @@ namespace Vänskap_Api.Service
             var friend = await _context.Users.SingleOrDefaultAsync(u => u.UserName == userName);
             if (friend == null) return false;
 
+            var isSent = await _context.FriendRequests.SingleOrDefaultAsync(u => u.SenderId == UserId && u.ReceiverId == friend.Id);
+            if (isSent != null) return false;
+
             var friendRequest = new FriendRequest()
             {
                 SenderId = UserId,
                 ReceiverId = friend.Id
             };
 
+            await _context.AddAsync(friendRequest);
+            await _context.SaveChangesAsync();
+
             return true;
         }
 
-        public async Task<(IEnumerable<string?> IncomingRequests, IEnumerable<string?> OutgoingRequests)> SeeFriendRequests()
+        public async Task<GetFriendRequestsDto> SeeFriendRequests()
         {
-            var incomingRequestTask = _context.FriendRequests
+            var incomingRequests = await _context.FriendRequests
                 .Where(f => f.ReceiverId == UserId)
                 .Include(f => f.Sender)
                 .Select(f => f.Sender != null ? f.Sender.UserName : "[Unknown sender]")
                 .ToListAsync();
             
-            var outgoingRequestTask = _context.FriendRequests
+            var outgoingRequests = await _context.FriendRequests
                 .Where(f => f.SenderId == UserId)
                 .Include(f => f.Receiver)
                 .Select(f => f.Receiver != null ? f.Receiver.UserName : "[Unknown receiver]")
                 .ToListAsync();
 
-            await Task.WhenAll(incomingRequestTask, outgoingRequestTask);
+            var friendsRequests = new GetFriendRequestsDto()
+            {
+                IncomingUserNames = incomingRequests!,
+                OutgoingUserNames = outgoingRequests!
+            };
 
-            return (incomingRequestTask.Result, outgoingRequestTask.Result);
+            return (friendsRequests);
         }
 
         public async Task<bool> AcceptFriendRequest(int id)
@@ -79,8 +90,17 @@ namespace Vänskap_Api.Service
                 var friendshipOne = new Friendship() { UserId = UserId, FriendId = friendRequest.SenderId };
                 var friendshipTwo = new Friendship() { UserId = friendRequest.SenderId, FriendId = UserId };
 
+                var user = await _context.Users.FindAsync(UserId);
+                var friend = await _context.Users.FindAsync(friendRequest.SenderId);
+
+                if (user == null || friend == null) return false;
+
+                user.Friendships.Add(friendshipOne);
+                friend.Friendships.Add(friendshipTwo);
+
                 _context.FriendRequests.Remove(friendRequest);
                 await _context.Friendships.AddRangeAsync(friendshipOne, friendshipTwo);
+
                 await _context.SaveChangesAsync();
                 return true;
             }
